+++ lib/nylas/api.rb
@@ -223,7 +223,8 @@ module Nylas
     def send!(message)
       return message.send! if message.respond_to?(:send!)
       return NewMessage.new(**message.merge(api: self)).send! if message.respond_to?(:key?)
-      return RawMessage.new(message, api: self).send! if message.is_a? String
+
+      RawMessage.new(message, api: self).send! if message.is_a? String
     end
 
     # Allows you to get an API that acts as a different user but otherwise has the same settings