+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ### Unreleased
 * Add support for count on message/thread search
+* Add support for detect provider endpoint
 
 ### 5.17.0 / 2022-04-04
 * Add support for verifying webhook signatures