+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 # Changelog
 
-### 5.17.0 / TBD
+### Unreleased
+* Add support for count on message/thread search
 * Add support for detect provider endpoint
 
 ### 5.17.0 / 2022-04-04