+++ CHANGELOG.md
@@ -1,5 +1,9 @@
 # Changelog
 
+### Unreleased
+* Add support for count on message/thread search
+* Add support for detect provider endpoint
+
 ### 5.17.0 / 2022-04-04
 * Add support for verifying webhook signatures
 * Add `event.updated_at`